--- components/function-controller/internal/controllers/serverless/gitops.go
@@ -1,1 +1,0 @@
-package serverless
